Big breakthough
Was able to complete excise the intersect/deintersect code as well as
code that was only used for building collision detection data. This
code was problematic because it used global variables that aliased
mutable references being used in the same part of the code. Rust does
not like this, and neither do I, and now I don't have to deal with it.